price python v0

+++ product_price_python/models/product_template.py
@@ -5,10 +5,10 @@ class ProductTemplate(models.Model):
 
     tipo_calculo        = fields.Selection([('personal', 'Servicios en horario habitual, extra y festivo')],
                                            string='Tipo', store=True, copy=True)
-    inicio_extra        = fields.Float(string="Inicio Extra")
-    inicio_ordinario    = fields.Float(string="Inicio Ordinaria")
-    final_ordinario     = fields.Float(string="Fin Ordinaria")
-    final_hextra        = fields.Float(string="Hora Extra límite")
+    inicio_extra        = fields.Integer(string="Inicio Extra")
+    inicio_ordinario    = fields.Integer(string="Inicio Ordinaria")
+    final_ordinario     = fields.Integer(string="Fin Ordinaria")
+    final_hextra        = fields.Integer(string="Hora Extra límite")
     hextra_factor       = fields.Float(string="Factor Hora Extra")
     hfestivo_factor     = fields.Float(string="Factor Festivo")
     horas_minimo        = fields.Integer(string="Horas mínimas")
